nextluk_mobile_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djongo==1.3.6
pymongo==4.6.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
"""
Django==4.2.0
mongoengine==0.27.0
pymongo==4.3.3
opencv-python==4.8.0.74
mediapipe==0.10.1
Pillow==9.5.0
numpy==1.24.3
scikit-learn==1.2.2
tensorflow==2.12.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-decouple==3.8
cloudinary==1.32.0
werkzeug==2.3.4
"""